Related tdf#99272: Prevent bogus com.sun.star.uno.Type(Short[].class) etc.
...so that the JNI UNO bridge will not accidentally reinterpret a Short[] as a
short[].  <https://wiki.openoffice.org/wiki/Uno/Java/Specifications/Type_Mapping>
makes it clear that the former is not a valid Java representation for UNO type
"sequence of short".

Needed one fix in bogus test code.  Also fixed the two bogus places in odk/examples
mentioned in f53e427291321eabe8d060a737e750a94739f911 "Resolves: tdf#99272 new
Short[] used instead of new short[]".

+++ odk/examples/DevelopersGuide/Forms/SalesFilter.java
@@ -86,7 +86,7 @@ class SalesFilter implements XActionListener, XPropertyChangeListener, XResetLis
             m_xFilterList.setPropertyValue( "Dropdown", Boolean.TRUE );
             m_xFilterList.setPropertyValue( "LineCount", Short.valueOf( (short)11 ) );
             m_xFilterList.setPropertyValue( "StringItemList", new String[] { "ever (means no filter)", "this morning", "1 week ago", "1 month ago", "1 year ago", "<other>" } );
-            m_xFilterList.setPropertyValue( "DefaultSelection", new Short[] { Short.valueOf( (short)0 ) } );
+            m_xFilterList.setPropertyValue( "DefaultSelection", new short[] { (short)0 } );
 
             m_xApplyFilter.setPropertyValue( "Label", "Apply Filter" );
